# packages/types/openapi.json
{
  "components": {
    "schemas": {
...
      "CabinetCreateIn": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          }
        },
        "required": [
          "name"
        ],
        "title": "CabinetCreateIn",
        "type": "object"
      },
      "CabinetMoveIn": {
        "description": "Reparent a cabinet to a different panel (same company).",
        "properties": {
          "panel_id": {
            "format": "uuid",
            "title": "Panel Id",
            "type": "string"
          }
        },
        "required": [
          "panel_id"
        ],
        "title": "CabinetMoveIn",
        "type": "object"
      },
      "CabinetOut": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "panel_id": {
            "format": "uuid",
            "title": "Panel Id",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "panel_id",
          "name",
          "position",
          "created_at",
          "updated_at"
        ],
        "title": "CabinetOut",
        "type": "object"
      },
      "CabinetUpdateIn": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          }
        },
        "title": "CabinetUpdateIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "PanelOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cabinets/{cabinet_id}": {
      "delete": {
        "operationId": "delete_cabinet_api_v1_cabinets__cabinet_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "cabinet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cabinet Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Company-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Company-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Cabinet",
        "tags": [
          "cabinets"
        ]
      },
      "put": {
        "operationId": "update_cabinet_api_v1_cabinets__cabinet_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "cabinet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cabinet Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Company-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Company-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CabinetUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CabinetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Cabinet",
        "tags": [
          "cabinets"
        ]
      }
    },
    "/api/v1/cabinets/{cabinet_id}/move": {
      "post": {
        "operationId": "move_cabinet_api_v1_cabinets__cabinet_id__move_post",
        "parameters": [
          {
            "in": "path",
            "name": "cabinet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cabinet Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Company-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Company-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CabinetMoveIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CabinetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Move Cabinet",
        "tags": [
          "cabinets"
        ]
      }
    },
...
    "/api/v1/panels/{panel_id}/cabinets": {
      "get": {
        "operationId": "list_cabinets_api_v1_panels__panel_id__cabinets_get",
        "parameters": [
          {
            "in": "path",
            "name": "panel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Panel Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Company-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Company-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CabinetOut"
                  },
                  "title": "Response List Cabinets Api V1 Panels  Panel Id  Cabinets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Cabinets",
        "tags": [
          "cabinets"
        ]
      },
      "post": {
        "operationId": "create_cabinet_api_v1_panels__panel_id__cabinets_post",
        "parameters": [
          {
            "in": "path",
            "name": "panel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Panel Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Company-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Company-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CabinetCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CabinetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Cabinet",
        "tags": [
          "cabinets"
        ]
      }
    },
...
  }
}